refactor(models): fold models/network.py into the models/asset/ package (#161)
The network shapes are all Amass OAM assets (AutonomousSystem via AsnRecord,
RIROrganization via RdapRecord, ContactRecord via Contact, DomainRecord), so
they belong in the OAM asset package alongside the rest.

- git mv models/network.py -> models/asset/network.py (history preserved);
  test mirror git mv models/test_network.py -> models/asset/test_network.py.
- Repointed all import sites: from models.network import X -> from
  models.asset.network import X (tools/recon/{rdap,asn}.py, models/asset/ip.py,
  and the recon / enrichment tests). models/asset/__init__ re-exports the
  network shapes, so from models.asset import AsnRecord works like the other
  asset shapes; from models import X is unchanged.
- ip.py now imports its AsnRecord / RdapRecord from the sibling
  models.asset.network; the package import order stays an acyclic DAG.

Pure relocation - no model shape or behaviour change.
