+++ tests/models/asset/test_network.py
